bditto-Django/requirements.txt
absl-py==0.11.0
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.1
astor==0.8.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
backcall==0.2.0
blis==0.4.1
cached-property==1.5.1
cachetools==4.2.1
catalogue==1.0.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.2
cloudinary==1.22.0
colorama==0.4.3
constantly==15.1.0
cryptography==3.2
cymem==2.0.3
daphne==2.5.0
decorator==4.4.2
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==2.2.10
django-admin-honeypot==1.1.0
django-allauth==0.42.0
django-appconf==1.0.3
django-braces==1.14.0
django-channels==0.7.0
django-common-helpers==0.9.2
django-cors-headers==3.2.0
django-countries==6.1.2
django-filter==2.3.0
django-heroku==0.3.1
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-oauth-toolkit==1.3.2
django-redis==4.12.1
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
easy-thumbnails==2.6
facebook-sdk==3.1.0
fastapi==0.45.0
future==0.18.2
gast==0.2.2
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
gunicorn==20.0.4
h11==0.9.0
h5py==3.1.0
hiredis==1.1.0
httpie==2.2.0
httplib2==0.18.1
httptools==0.1.1
hyperlink==20.0.1
idna==2.8
importlib-metadata==1.7.0
incremental==17.5.0
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==0.16.0
jsonfield==3.1.0
jupyter-client==6.1.11
jupyter-core==4.7.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
more-itertools==8.4.0
msgpack==0.6.2
murmurhash==1.0.2
numpy==1.16.0
oauthlib==3.1.0
opt-einsum==3.3.0
ordered-set==3.1.1
ordered-set-stubs==0.1.3
pandas==1.1.5
parso==0.8.1
pdfkit==0.6.1
pickleshare==0.7.5
Pillow==6.2.1
plac==1.1.3
poetry-version==0.1.5
preshed==3.0.2
prompt-toolkit==3.0.14
protobuf==3.14.0
psycopg2==2.8.4
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.6.1
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-decouple==3.3
python-http-client==3.2.7
python-twitter==3.5
python3-openid==3.2.0
pytz==2019.3
pywin32==300
pyzmq==22.0.2
redis==3.5.3
regex==2020.7.14
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.7
ruamel.yaml==0.15.100
scipy==1.5.2
sendgrid==6.4.1
service-identity==18.1.0
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sqlparse==0.3.0
srsly==1.0.2
starkbank-ecdsa==1.0.0
starlette==0.12.9
swapper==1.1.2.post1
tb-nightly==1.14.0a20190603
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-hub==0.10.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
thinc==7.4.1
threadpoolctl==2.1.0
tomlkit==0.5.11
tornado==6.1
tqdm==4.48.0
traitlets==5.0.5
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.7
uvicorn==0.11.6
wasabi==0.7.1
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
whitenoise==4.1.4
wkhtmltopdf==0.2
wrapt==1.12.1
xlrd==2.0.1
zipp==3.1.0
zope.interface==5.2.0
